期初建账

+++ InvoicingSystemServer/src/dataservice/logdataservice/LogDataService.java
@@ -16,5 +16,5 @@ public interface LogDataService extends Remote{
 	public int getLogNum() throws RemoteException;
 	public boolean delete(Date time) throws RemoteException;
 	public boolean update(LogPO po) throws RemoteException;
-	public int getLargestIDofLog() throws RemoteException;
+	public long getLargestIDofLog() throws RemoteException;
 }
